requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
pydantic>=2.0.0
aiosqlite>=0.19.0
